==> scripts/5_Loading_MongoDB.ipynb <==
import pandas as pd
from pymongo import MongoClient
from pathlib import Path

base_dir = Path().resolve().parent
archivo_csv = base_dir / "database" / "pacientes_clean.csv"

df = pd.read_csv(archivo_csv)
print(f"Registros a insertar: {len(df)}")

cliente = MongoClient("mongodb://localhost:27017/")

db = cliente["SaludDigital_2025"]
coleccion = db["pacientes"]

coleccion.delete_many({})

registros = df.to_dict(orient="records")

coleccion.insert_many(registros)

print("Inserción completada correctamente en MongoDB.")
print(f"Total de documentos en la colección: {coleccion.count_documents({})}")
